refactor(ui): rename IAnnouncements to IAnnouncement
