+++ ui/admin/src/interfaces/IAnnouncement.ts
@@ -1,0 +1,3 @@
+import { IAnnouncementShort } from "@/interfaces/IAnnouncement";
+
+export type IAdminAnnouncementShort = IAnnouncementShort;
