--- ui/admin/src/interfaces/IAnnouncements.ts
@@ -1,3 +1,0 @@
-import { IAnnouncementShort } from "@/interfaces/IAnnouncements";
-
-export type IAdminAnnouncementShort = IAnnouncementShort;
