+++ ui/src/interfaces/IAnnouncement.ts
